+++ src/tobac_flow/utils/__init__.py
@@ -6,6 +6,7 @@ General purpose functions which do not depend on other modules in the package
 
 from tobac_flow.utils.datetime_utils import *
 from tobac_flow.utils.filter_utils import *
+from tobac_flow.utils.flow_utils import *
 from tobac_flow.utils.geo_utils import *
 from tobac_flow.utils.label_utils import *
 from tobac_flow.utils.normalisation_utils import *
